chore: updated to rust 1.69.0, fixed associated warnings
