+++ rust-toolchain.toml
@@ -1,4 +1,10 @@
 [toolchain]
-channel = "1.68.2"
+channel = "1.69.0"
 components = ["rustfmt", "clippy"]
-targets = ["wasm32-unknown-unknown", "wasm32-wasi", "aarch64-unknown-linux-gnu", "x86_64-pc-windows-gnu", "aarch64-apple-darwin"]
+targets = [
+  "wasm32-unknown-unknown",
+  "wasm32-wasi",
+  "aarch64-unknown-linux-gnu",
+  "x86_64-pc-windows-gnu",
+  "aarch64-apple-darwin"
+]
