+++ templates/rust/Cargo.toml
@@ -20,7 +20,7 @@ wick-component = { git = "https://github.com/candlecorp/wick.git" }
 wasmrs-guest = { version = "0.11", features = [] }
 serde = { version = "1", features = ["derive"] }
 async-trait = "0.1"
-anyhow = { version = "1.0" }
+anyhow = { version = "1" }
 
 [build-dependencies]
 wick-component-codegen = { git = "https://github.com/candlecorp/wick.git" }
